Switch module ids to u8 from u32

+++ wfc_dylib/ffi/wfc.h
@@ -33,8 +33,8 @@ typedef WfcState *Wfc;
 
 typedef struct {
   AdjacencyRuleKind kind;
-  uint32_t module_low;
-  uint32_t module_high;
+  uint8_t module_low;
+  uint8_t module_high;
 } AdjacencyRule;
 
 /**
